Backend/requirements.txt
fastapi
uvicorn
pypdf2
pytesseract
pdf2image
pillow
langchain_groq
